factor out common cache subcomponents into uncore.util
