+++ src/main/scala/coreplex/Configs.scala
@@ -11,6 +11,7 @@ import uncore.coherence._
 import uncore.agents._
 import uncore.devices._
 import uncore.converters._
+import uncore.util._
 import rocket._
 import util._
 import util.ConfigUtils._
