+++ src/main/scala/groundtest/Configs.scala
@@ -6,6 +6,7 @@ import diplomacy._
 import uncore.tilelink._
 import uncore.coherence._
 import uncore.agents._
+import uncore.util._
 import uncore.devices.NTiles
 import junctions._
 import config._
